MinGit fixes

+++ install_Applio.bat
@@ -25,6 +25,7 @@ set "CONDA_ROOT_PREFIX=%UserProfile%\Miniconda3"
 set "INSTALL_ENV_DIR=%principal%\env"
 set "MINICONDA_DOWNLOAD_URL=https://repo.anaconda.com/miniconda/Miniconda3-py39_23.9.0-0-Windows-x86_64.exe"
 set "CONDA_EXECUTABLE=%CONDA_ROOT_PREFIX%\Scripts\conda.exe"
+set "tempFile=%cd%\powershell_output.txt"
 
 echo.
 cls
@@ -37,34 +38,72 @@ cls
 for /f "delims=: tokens=*" %%A in ('findstr /b ":::" "%~f0"') do @echo(%%A
 echo.
 
+if not exist "%cd%\mingit.zip" (
 echo Downloading MinGit from %URL_EXTRA%/mingit.zip
 curl -s -LJO %URL_EXTRA%/mingit.zip -o mingit.zip
+)
 
 if not exist "%cd%\mingit.zip" (
 echo Download failed trying with the powershell method
 powershell -Command "& {Invoke-WebRequest -Uri '%URL_EXTRA%/mingit.zip' -OutFile 'mingit.zip'}"
 )
 
-set "powershellCommand=try { Add-Type -AssemblyName 'System.IO.Compression.FileSystem'; [System.IO.Compression.ZipFile]::ExtractToDirectory('%cd%\mingit.zip', '%cd%'); goto goodExtract } catch { goto errorExtract }"
-powershell -command "& {%powershellCommand%}"
+set powershellOutput="goodExtract"
 
-:errorExtract
-echo Error extracting! Deleting file and trying directly with powershell method!
-del mingit.zip
-powershell -Command "& {Invoke-WebRequest -Uri '%URL_EXTRA%/mingit.zip' -OutFile 'mingit.zip'}"
-powershell -command "& {%powershellCommand%}"
+if not exist "%cd%\mingit" (
+    set "powershellScript=try { Add-Type -AssemblyName 'System.IO.Compression.FileSystem'; [System.IO.Compression.ZipFile]::ExtractToDirectory('%cd%\mingit.zip', '%cd%'); Write-Output 'goodExtract' } catch { Write-Output 'errorExtract' }"
 
-:goodExtract
-echo Cloning the repository...
-%cd%\mingit\cmd\git.exe clone %repoUrl% %repoFolder%
-echo Moving the mingit folder...
-robocopy "%cd%\mingit" "%principal%\lib\tools\mingit" /e /move /dcopy:t > NUL
-if errorlevel 8 echo Warnings or errors occurred during the move.
-cd %repoFolder%
-del install_Applio.bat
-del /q *.sh
-echo.
-cls
+    powershell -NoProfile -ExecutionPolicy Bypass -Command "%powershellScript%" > "%tempFile%"
+
+    set /p powershellOutput=<"%tempFile%"
+
+    del "%tempFile%"
+)
+
+if "%powershellOutput%"=="goodExtract" (
+    del mingit.zip
+    echo Cloning the repository...
+    %cd%\mingit\cmd\git.exe clone %repoUrl% %repoFolder%
+    pause
+    echo Moving the mingit folder...
+    pause
+    robocopy "%cd%\mingit" "%principal%\lib\tools\mingit" /e /move /dcopy:t > NUL
+    if errorlevel 8 echo Warnings or errors occurred during the move.
+    pause
+    cd %repoFolder%
+    del install_Applio.bat
+    del /q *.sh
+    echo.
+    cls
+) else (
+    echo Error extracting! Deleting file and trying directly with PowerShell method!
+    del mingit.zip
+    powershell -Command "& {Invoke-WebRequest -Uri '%URL_EXTRA%/mingit.zip' -OutFile 'mingit.zip'}"
+    powershell -NoProfile -ExecutionPolicy Bypass -Command "%powershellScript%" > "%tempFile%"
+    set /p powershellOutput=<"%tempFile%"
+    del "%tempFile%"
+    if "%powershellOutput%"=="errorExtract" (
+        echo Theres a problem extracting the file please download the file and extract it manually 
+        echo https://huggingface.co/IAHispano/applio/resolve/main/mingit.zip
+        pause
+        exit
+    ) else (
+        del mingit.zip
+        echo Cloning the repository...
+        %cd%\mingit\cmd\git.exe clone %repoUrl% %repoFolder%
+        pause
+        echo Moving the mingit folder...
+        robocopy "%cd%\mingit" "%principal%\lib\tools\mingit" /e /move /dcopy:t > NUL
+        if errorlevel 8 echo Warnings or errors occurred during the move.
+        pause
+        cd %repoFolder%
+        del install_Applio.bat
+        del /q *.sh
+        pause
+        echo.
+        cls
+    )
+)
 
 :menu
 echo Installing dependencies...
